Switches example to wazero and introduces tags for CGO engines (#30)
The CGO implementations (wasmer and wasmtime) clash on how they
implement WASI due to a flat namespace. This adjusts the build tags
accordingly.

This also consolidates the example code which drifted recently. Before,
it used different code in the README and the example/main.go and even
different runtimes. Instead, this uses wazero which has no source of
conflicts, but also documents how to change the runtime to any supported
engine.

Fixes #26

Signed-off-by: Adrian Cole <adrian@tetrate.io>